move errorResponseFrom function to common
